[#3171] - use BigInteger and add tests

+++ hs_core/migrations/0041_baseresource__size.py
@@ -15,6 +15,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='baseresource',
             name='_size',
-            field=models.IntegerField(default=0),
+            field=models.BigIntegerField(default=0),
         ),
     ]
